If database exists, create_test_environment.sh now prompts for db reset

+++ doc/create_test_environment.sh
@@ -6,37 +6,36 @@ cd ..
 
 VENV_DIR=../aplusenv
 VENV_PYTHON=$VENV_DIR/bin/python
-FIXTURE_DIR=course/fixtures
 
 
-#remove old test environment
+# (re)create test environment
 if [ -d $VENV_DIR ]; then
     rm -R $VENV_DIR
 fi
-if [ -f aplus.db ]; then
-    rm aplus.db
-fi
-mkdir -p $FIXTURE_DIR
-if [ -f $FIXTURE_DIR/initial_data.json ]; then
-    rm $FIXTURE_DIR/initial_data.json
-fi
-
-
-# create the virtualenv
 python venv_bootstrap.py $VENV_DIR
 
+# (re)create the database
+if [ -f aplus.db ]; then
+    while true; do
+        read -p "Do you wish to reset the database as well (Y/N)?" yn
+        case $yn in
+            [Yy]*)
+                rm aplus.db
+                $VENV_PYTHON manage.py syncdb --noinput
+                $VENV_PYTHON manage.py migrate
+                $VENV_PYTHON manage.py loaddata doc/initial_data.json
+                $VENV_PYTHON manage.py createsuperuser
+                break
+                ;;
+            [Nn]*)
+                exit
+                ;;
+            *)
+                echo "Invalid option! Please answer Y (yes) or N (no)"
+                ;;
+        esac
+    done
+fi
 
-# create the database
-$VENV_PYTHON manage.py syncdb --noinput
-$VENV_PYTHON manage.py migrate
-
-
-#insert test data
-cp doc/initial_data.json $FIXTURE_DIR/initial_data.json
-$VENV_PYTHON manage.py migrate course
-
-
-#create super user
-$VENV_PYTHON manage.py createsuperuser
